Switched to header and source file structure.
Reformatted some files and started on tests.
