+++ cpp_src/src/astar.cpp
@@ -1,34 +1,9 @@
 // Std includes
-#include <iostream>
+#include <unordered_map>
 #include <queue>
 
 // Custom includes
-#include "primitives.h"
-
-// ----- CONSTANTS ------------------------------
-/* Represents the north, south, east, west, north-east, north-west, south-east
- * and south-west directions on a compass */
-const std::vector<Point> INTERCARDINAL_OFFSETS = {
-    {-1, -1}, {0, -1}, {1, -1}, {-1, 0}, {1, 0}, {-1, 1}, {0, 1}, {1, 1},
-};
-
-// ----- STRUCTURES ------------------------------
-/// Represents a grid position and its costs from the start position
-///
-/// Parameters
-/// ----------
-/// cost - The cost to traverse to this neighbour.
-/// pair - The position in the grid.
-struct Neighbour {
-  int cost;
-  Point pair;
-
-  inline bool operator<(const Neighbour nghbr) const {
-    // The priority_queue data structure gets the maximum priority, so we need
-    // to override that functionality to get the minimum priority
-    return cost > nghbr.cost;
-  }
-};
+#include "astar.hpp"
 
 // ----- FUNCTIONS ------------------------------
 /// Get a target's neighbours based on a given list of offsets.
@@ -83,8 +58,8 @@ calculate_astar_path(std::vector<std::vector<TileType>> &grid, Point &start,
   std::priority_queue<Neighbour> queue;
   std::unordered_map<Point, Point> came_from = {{start, start}};
   std::unordered_map<Point, int> distances = {{start, 0}};
-  int height = (int)grid.capacity();
-  int width = (int)grid[0].capacity();
+  int height = (int) grid.capacity();
+  int width = (int) grid[0].capacity();
   queue.push({0, start});
 
   // Loop until the priority queue is empty
@@ -126,9 +101,9 @@ calculate_astar_path(std::vector<std::vector<TileType>> &grid, Point &start,
         // infinity, otherwise, set it to f = g + h
         int f_cost =
             (grid[neighbour.y][neighbour.x] == TileType::Obstacle)
-                ? std::numeric_limits<int>::max()
-                : distances[neighbour] + std::max(abs(neighbour.x - current.x),
-                                                  abs(neighbour.y - current.y));
+            ? std::numeric_limits<int>::max()
+            : distances[neighbour] + std::max(abs(neighbour.x - current.x),
+                                              abs(neighbour.y - current.y));
 
         // Add the neighbour to the priority queue
         queue.push({f_cost, neighbour});
