+++ cpp_src/test/test_map.cpp
@@ -1,0 +1,5 @@
+// External includes
+#include "gtest/gtest.h"
+
+// Custom includes
+#include "map.hpp"
